chore(Posthog): move sdk loader to different file
