+++ src/integrations/Posthog/browser.js
@@ -3,7 +3,6 @@
 import get from 'get-value';
 import logger from '../../utils/logUtil';
 import { removeTrailingSlashes } from '../../utils/utils';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
 
 class Posthog {
@@ -22,7 +21,8 @@ class Posthog {
     this.propertyBlackList = [];
     this.xhrHeaders = {};
     this.enableLocalStoragePersistence = config.enableLocalStoragePersistence;
-    this.areTransformationsConnected = destinationInfo && destinationInfo.areTransformationsConnected;
+    this.areTransformationsConnected =
+      destinationInfo && destinationInfo.areTransformationsConnected;
     this.destinationId = destinationInfo && destinationInfo.destinationId;
 
     if (config.xhrHeaders && config.xhrHeaders.length > 0) {
@@ -53,48 +53,7 @@ class Posthog {
       logger.debug('===[POSTHOG]: loadIntegration flag is disabled===');
       return;
     }
-    !(function (t, e) {
-      var o, n, p, r;
-      e.__SV ||
-        ((window.posthog = e),
-        (e._i = []),
-        (e.init = function (i, s, a) {
-          function g(t, e) {
-            var o = e.split('.');
-            2 == o.length && ((t = t[o[0]]), (e = o[1])),
-              (t[e] = function () {
-                t.push([e].concat(Array.prototype.slice.call(arguments, 0)));
-              });
-          }
-          ((p = t.createElement('script')).type = 'text/javascript'),
-            (p.async = !0),
-            p.setAttribute('data-loader', LOAD_ORIGIN),
-            (p.src = s.api_host + '/static/array.js'),
-            (r = t.getElementsByTagName('script')[0]).parentNode.insertBefore(p, r);
-          var u = e;
-          for (
-            void 0 !== a ? (u = e[a] = []) : (a = 'posthog'),
-              u.people = u.people || [],
-              u.toString = function (t) {
-                var e = 'posthog';
-                return 'posthog' !== a && (e += '.' + a), t || (e += ' (stub)'), e;
-              },
-              u.people.toString = function () {
-                return u.toString(1) + '.people (stub)';
-              },
-              o =
-                'capture identify alias people.set people.set_once set_config register register_once unregister opt_out_capturing has_opted_out_capturing opt_in_capturing reset isFeatureEnabled onFeatureFlags'.split(
-                  ' ',
-                ),
-              n = 0;
-            n < o.length;
-            n++
-          )
-            g(u, o[n]);
-          e._i.push([i, s, a]);
-        }),
-        (e.__SV = 1));
-    })(document, window.posthog || []);
+    loader(document, window.posthog || []);
 
     const configObject = {
       api_host: this.yourInstance,
